Refactor useStaticQuery
